box drawing initial implementation

+++ src/gfx/texture_atlas.cpp
@@ -8,13 +8,13 @@ using namespace wgpu;
 TextureAtlas::TextureAtlas(float _glyphSize, float _dpiScale)
     : dpiScale(_dpiScale), trueGlyphSize(_glyphSize * dpiScale) {
 
-  uint initialTextureHeight = trueGlyphSize * 3;
+  int initialTextureHeight = trueGlyphSize * 3;
   bufferSize = {trueGlyphSize * glyphsPerRow, initialTextureHeight};
   textureSize = glm::vec2(bufferSize) / dpiScale;
 
   dataRaw.resize(bufferSize.x * bufferSize.y);
   data =
-    std::mdspan(dataRaw.data(), std::dextents<uint, 2>{bufferSize.y, bufferSize.x});
+    std::mdspan(dataRaw.data(), std::dextents<size_t, 2>{bufferSize.y, bufferSize.x});
 
   // init bind group data
   textureSizeBuffer =
@@ -30,7 +30,7 @@ TextureAtlas::TextureAtlas(float _glyphSize, float _dpiScale)
   renderTexture = RenderTexture(textureSize, dpiScale, TextureFormat::RGBA8Unorm);
 }
 
-Region TextureAtlas::AddGlyph(std::mdspan<uint8_t, std::dextents<uint, 2>> glyphData) {
+Region TextureAtlas::AddGlyph(std::mdspan<uint8_t, std::dextents<size_t, 2>> glyphData) {
   // check if current row is full
   // if so, move to next row
   if (currentPos.x + glyphData.extent(1) > bufferSize.x) {
@@ -43,8 +43,8 @@ Region TextureAtlas::AddGlyph(std::mdspan<uint8_t, std::dextents<uint, 2>> glyph
   }
 
   // fill data
-  for (uint row = 0; row < glyphData.extent(0); row++) {
-    for (uint col = 0; col < glyphData.extent(1); col++) {
+  for (size_t row = 0; row < glyphData.extent(0); row++) {
+    for (size_t col = 0; col < glyphData.extent(1); col++) {
       Color& dest = data[currentPos.y + row, currentPos.x + col];
       dest.r = 255;
       dest.g = 255;
@@ -60,21 +60,22 @@ Region TextureAtlas::AddGlyph(std::mdspan<uint8_t, std::dextents<uint, 2>> glyph
 
   // advance current position
   currentPos.x += glyphData.extent(1);
-  currMaxHeight = std::max(currMaxHeight, glyphData.extent(0));
+  currMaxHeight = std::max((size_t)currMaxHeight, glyphData.extent(0));
 
   return MakeRegion(regionPos, regionSize);
 }
 
 void TextureAtlas::Resize() {
-  uint heightIncrease = trueGlyphSize * 3;
+  int heightIncrease = trueGlyphSize * 3;
   bufferSize.y += heightIncrease;
   textureSize = glm::vec2(bufferSize) / dpiScale;
 
   dataRaw.resize(bufferSize.x * bufferSize.y);
   data =
-    std::mdspan(dataRaw.data(), std::dextents<uint, 2>{bufferSize.y, bufferSize.x});
+    std::mdspan(dataRaw.data(), std::dextents<size_t, 2>{bufferSize.y, bufferSize.x});
 
   resized = true;
+  // LOG_INFO("Resized texture atlas to {}x{}", bufferSize.x, bufferSize.y);
 }
 
 void TextureAtlas::Update() {
